+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
             android:value="ca-app-pub-3940256099942544~3347511713">
         </meta-data>
 
-        <activity android:name=".MainActivity"
+        <activity android:name=".ui.MainActivity"
             android:configChanges="layoutDirection|orientation|screenSize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
