+++ app/src/main/java/com/android/kpopdance/data/Youtube.kt
@@ -1,4 +1,4 @@
-package com.android.kpopdance
+package com.android.kpopdance.data
 
 data class Youtube (
     val id: String,
