+++ app/src/main/java/com/android/kpopdance/ui/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.android.kpopdance
+package com.android.kpopdance.ui
 
 import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
@@ -7,6 +7,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.android.kpopdance.viewmodel.HomeViewModel
+import com.android.kpopdance.R
+import com.android.kpopdance.data.Youtube
 import kotlinx.android.synthetic.main.home_fragment.*
 
 
@@ -50,24 +53,24 @@ class HomeFragment : Fragment() {
                 "2021.01.16"
             ),
             Youtube(
-                    "N-n3hEJKvC4",
-                    "(G)I-DLE - HWAA Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "N-n3hEJKvC4",
+                "(G)I-DLE - HWAA Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "FPjbtai9rx0",
-                    "aespa (에스파) - Black Mamba Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "FPjbtai9rx0",
+                "aespa (에스파) - Black Mamba Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "nvKDrmGP4Is",
-                    "TWICE - I CAN'T STOP ME Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "nvKDrmGP4Is",
+                "TWICE - I CAN'T STOP ME Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "RDFJQmNCBoU",
-                    "ITZY - Not Shy Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "RDFJQmNCBoU",
+                "ITZY - Not Shy Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
                 "ogGuOZ_6O5c",
