+++ app/src/main/java/com/android/kpopdance/ui/SearchFragment.kt
@@ -1,4 +1,4 @@
-package com.android.kpopdance
+package com.android.kpopdance.ui
 
 import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
@@ -6,6 +6,9 @@ import android.view.*
 import android.widget.ImageButton
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.android.kpopdance.R
+import com.android.kpopdance.viewmodel.SearchViewModel
+import com.android.kpopdance.data.Youtube
 import com.google.android.gms.ads.AdRequest
 import kotlinx.android.synthetic.main.search_fragment.*
 
@@ -32,49 +35,49 @@ class SearchFragment : Fragment() {
         // for test
         val youtubes = arrayListOf(
             Youtube(
-                    "N-n3hEJKvC4",
-                    "(G)I-DLE - HWAA Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "N-n3hEJKvC4",
+                "(G)I-DLE - HWAA Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "FPjbtai9rx0",
-                    "aespa (에스파) - Black Mamba Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "FPjbtai9rx0",
+                "aespa (에스파) - Black Mamba Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "nvKDrmGP4Is",
-                    "TWICE - I CAN'T STOP ME Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "nvKDrmGP4Is",
+                "TWICE - I CAN'T STOP ME Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "RDFJQmNCBoU",
-                    "ITZY - Not Shy Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "RDFJQmNCBoU",
+                "ITZY - Not Shy Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "N-n3hEJKvC4",
-                    "(G)I-DLE - HWAA Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "N-n3hEJKvC4",
+                "(G)I-DLE - HWAA Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "FPjbtai9rx0",
-                    "aespa (에스파) - Black Mamba Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "FPjbtai9rx0",
+                "aespa (에스파) - Black Mamba Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "nvKDrmGP4Is",
-                    "TWICE - I CAN'T STOP ME Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "nvKDrmGP4Is",
+                "TWICE - I CAN'T STOP ME Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "RDFJQmNCBoU",
-                    "ITZY - Not Shy Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "RDFJQmNCBoU",
+                "ITZY - Not Shy Dance Practice (Mirrored)",
+                "2021.01.16"
             ),
             Youtube(
-                    "ogGuOZ_6O5c",
-                    "CHUNG HA (청하) - Roller Coaster Dance Practice (Mirrored)",
-                    "2021.01.16"
+                "ogGuOZ_6O5c",
+                "CHUNG HA (청하) - Roller Coaster Dance Practice (Mirrored)",
+                "2021.01.16"
             )
         )
