+++ app/src/main/java/com/android/kpopdance/ui/YoutubeBigAdapter.kt
@@ -1,4 +1,4 @@
-package com.android.kpopdance
+package com.android.kpopdance.ui
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,6 +6,8 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.android.kpopdance.R
+import com.android.kpopdance.data.Youtube
 import com.bumptech.glide.Glide
 
 class YoutubeBigAdapter : RecyclerView.Adapter<YoutubeBigAdapter.ViewHolder>() {
