+++ app/src/main/res/layout/search_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchFragment">
+    tools:context=".ui.SearchFragment">
 
     <androidx.drawerlayout.widget.DrawerLayout
         android:id="@+id/drawer"
