+++ app/src/main/res/layout/youtube_small_item.xml
@@ -27,12 +27,11 @@
             android:id="@+id/titleTextView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:ellipsize="end"
+            android:maxLines="2"
             android:text="TextView"
             android:textColor="#000000"
-            android:textColorHint="#00000000"
-            android:textSize="14sp"
-            android:maxLines="2"
-            android:ellipsize="end"/>
+            android:textSize="14sp" />
 
         <LinearLayout
             android:layout_width="match_parent"
